Refactored Abstractions
