+++ API/Landlord/Controllers/LandlordController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Abstractions;
-using Abstractions.Models;
+using Abstractions.Models.Results;
 using Abstractions.Repositories;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
