+++ Abstractions/Models/Issue.cs
@@ -1,11 +1,11 @@
 ﻿namespace Abstractions.Models
 {
-    public interface IIssue
+    public class Issue
     {
         //  Properties
         //  ==========
 
         public int Id { get; set; }
-        public string Content { get; set; }
+        public string Content { get; set; } = null!;
     }
 }
