+++ Abstractions/Repositories/IIssueRepository.cs
@@ -9,8 +9,8 @@ namespace Abstractions.Repositories
         //  Methods
         //  =======
 
-        public Task<IEnumerable<IIssue>> GetAllIssues();
-        public Task<IIssue> GetIssueById(int id);
+        public Task<IEnumerable<Issue>> GetAllIssues();
+        public Task<Issue> GetIssueById(int id);
         public Task<bool> CreateIssue(string content);
     }
 }
