+++ SQLServer/Models/RegisterTenantResult.cs
@@ -1,4 +1,4 @@
-﻿using Abstractions.Models;
+﻿using Abstractions.Models.Results;
 using System;
 using System.Collections.Generic;
 using System.Text;
